# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Levels of Rio Negro - AM",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/levels/{token}": {
      "get": {
        "tags": [
          "Levels"
        ],
        "description": "Get Levels of Rio Negro",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/sign-in": {
      "post": {
        "tags": [
          "Acess User"
        ],
        "description": "Create token acess.",
        "parameters": [
          {
            "name": "token",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUserToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sign-out/{token}": {
      "delete": {
        "tags": [
          "Acess User"
        ],
        "description": "Delete token acess.",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "AddUserToken": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "token": {
          "type": "string",
          "example": "1b9d6bcd-bbfd-4b2d-9b5d-ab8dfbbd4bed"
        }
      },
      "required": [
        "name",
        "token"
      ]
    }
  }
}